Ionic Framework Textarea issue
I think this project shows the wrong way, or a bug, to do textures on
ionic framework.
If you tap on the text area (especially on an iPad Retina), the
textarea will get focus and the keyboard will come up.
You can swipe the text area and it will expose a white background.
grabbing other elements will allow you drag them independent of the
background.